requirements.txt
fastapi
uvicorn
line-bot-sdk
python-dotenv
supabase
streamlit
pandas
google-genai
spacy
ja-ginza